+++ internal/ent/githubflow_create.go
@@ -26,7 +26,7 @@ import (
 
 	"entgo.io/ent/dialect/sql/sqlgraph"
 	"entgo.io/ent/schema/field"
-	"github.com/loopholelabs/auth/ent/githubflow"
+	"github.com/loopholelabs/auth/internal/ent/githubflow"
 )
 
 // GithubFlowCreate is the builder for creating a GithubFlow entity.
