+++ pkg/provider/github/database.go
@@ -18,7 +18,7 @@ package github
 
 import (
 	"context"
-	"github.com/loopholelabs/auth/ent"
+	"github.com/loopholelabs/auth/internal/ent"
 	"time"
 )
